refactor: 세션 삭제 스레드 세팅 위치 변경
