+++ tomcat/src/main/java/org/apache/catalina/connector/Connector.java
@@ -23,7 +23,6 @@ public class Connector implements Runnable {
     private static final int DEFAULT_PORT = 8080;
     private static final int DEFAULT_ACCEPT_COUNT = 350;
     private static final int DEFAULT_THREAD_POOL_SIZE = 250;
-    private static final int SESSION_EXPIRED_TIME = 30;
 
     private final ServerSocket serverSocket;
     private boolean stopped;
@@ -48,17 +47,6 @@ public class Connector implements Runnable {
                 timeUnit,
                 new LinkedBlockingQueue<>()
         );
-        settingRemoveSessionThread();
-    }
-
-    private void settingRemoveSessionThread() {
-        ScheduledExecutorService sessionRemover = Executors.newScheduledThreadPool(1);
-        Runnable runnable = () -> {
-            for (final Session session : SessionManager.findExpiredSession()) {
-                SessionManager.remove(session);
-            }
-        };
-        sessionRemover.scheduleAtFixedRate(runnable, 1, SESSION_EXPIRED_TIME, TimeUnit.MINUTES);
     }
 
     private ServerSocket createServerSocket(final int port, final int acceptCount) {
