+++ tomcat/src/main/java/org/apache/coyote/http/SessionManager.java
@@ -7,10 +7,28 @@ import java.util.Map.Entry;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.TimeUnit;
 
 public class SessionManager {
 
+    private static final int SESSION_EXPIRED_TIME = 30;
     private static final Map<String, Session> SESSIONS = new ConcurrentHashMap<>();
+
+    public SessionManager(){
+        settingRemoveSessionThread();
+    }
+
+    private void settingRemoveSessionThread() {
+        ScheduledExecutorService sessionRemover = Executors.newScheduledThreadPool(1);
+        Runnable runnable = () -> {
+            for (final Session session : findExpiredSession()) {
+                remove(session);
+            }
+        };
+        sessionRemover.scheduleAtFixedRate(runnable, 1, SESSION_EXPIRED_TIME, TimeUnit.MINUTES);
+    }
 
     public static Optional<Session> add() {
         final Session session = new Session(UUID.randomUUID().toString());
@@ -34,8 +52,5 @@ public class SessionManager {
 
     public static void remove(Session session) {
         SESSIONS.remove(session);
-    }
-
-    public SessionManager() {
     }
 }
